Checkpoint compiling native interfaces
