+++ packages/react-native-v2/ios/EvervaultPaymentViewManager.mm
@@ -1,6 +1,6 @@
 #import <React/RCTViewManager.h>
 
-@interface RCT_EXTERN_MODULE(EvervaultPaymentViewManager, RCTViewManager)
+@interface RCT_EXTERN_MODULE(ApplePayButton, RCTViewManager)
 
 RCT_EXPORT_VIEW_PROPERTY(config, NSDictionary)
 RCT_EXPORT_VIEW_PROPERTY(transaction, NSDictionary)
