+++ packages/react-native-v2/src/ApplePayButton/types.tsx
@@ -1,107 +1,10 @@
-import { type ViewProps } from 'react-native';
-import type { DirectEventHandler, Double, WithDefault } from 'react-native/Libraries/Types/CodegenTypes';
-
-export class Amount {
-  public value: string;
-
-  constructor(value: string) {
-    this.value = value;
-  }
-};
-
-export type SummaryItem = {
-  label: string;
-  amount: Amount;
-};
-
-export type OneOffTransaction = {
-  currency: string;
-  country: string;
-  paymentSummaryItems: SummaryItem[];
-  // TODO: Add other fields
-}
-
-export type RecurringTransaction = {
-  // TODO: Add other fields
-};
-
-export type DisbursementTransaction = {
-  // TODO: Add other fields
-};
-
-export type Transaction = OneOffTransaction | RecurringTransaction | DisbursementTransaction;
-
-export type Config = {
-  appId: string;
-  merchantId: string;
-  supportedNetworks: CardNetwork[];
-  buttonType: string;
-  buttonTheme: string;
-};
-
-export type ApplePayNetworkTokenExpiry = {
-  month: string;
-  year: string;
-}
-
-export type ApplePayNetworkToken = {
-  number: string;
-  expiry: ApplePayNetworkTokenExpiry;
-  rawExpiry: string;
-  tokenServiceProvider: string;
-};
-
-export type ApplePayCard = {
-  brand?: string;
-  funding?: string;
-  segment?: string;
-  country?: string;
-  currency?: string;
-  issuer?: string;
-};
-
-export type ApplePayResponse = {
-  networkToken: ApplePayNetworkToken;
-  card: ApplePayCard;
-  cryptogram: string;
-  eci?: string;
-  paymentDataType: string;
-  deviceManufacturerIdentifier: string;
-}
+import { ApplePayResponse, CardNetwork, Config, Transaction } from '../specs/ApplePayButtonNativeComponent';
 
 export type ButtonType = 'plain' | 'book' | 'buy' | 'checkout' | 'order' | 'subscribe' | 'pay' | 'in_store' | 'donate' | 'reload' | 'add_money' | 'top_up' | 'rent' | 'support' | 'contribute' | 'tip' | 'continue';
 export type ButtonTheme = 'automatic' | 'white' | 'white_outline' | 'black';
 export type AuthMethod = 'PAN_ONLY' | 'CRYPTOGRAM_3DS';
-export type CardNetwork = 'visa' | 'mastercard' | 'amex' | 'discover' | 'jcb';
 
-export interface NativeProps extends ViewProps {
-  config: Config;
-  transaction: Transaction;
-  onDidAuthorizePayment?: DirectEventHandler<{
-    networkToken: {
-      number: string;
-      expiry: { month: string; year: string };
-      rawExpiry: string;
-      tokenServiceProvider: string;
-    };
-    card: {
-      brand?: string;
-      funding?: string;
-      segment?: string;
-      country?: string;
-      currency?: string;
-      issuer?: string;
-    };
-    cryptogram: string;
-    eci?: string;
-    paymentDataType: string;
-    deviceManufacturerIdentifier: string;
-  }>;
-  onDidFinishWithResult?: DirectEventHandler<{ success: boolean; error?: string }>;
-  onPrepareTransaction?: DirectEventHandler<null>;
-}
-
-export type ApplePayButtonProps = Config & {
+export type ApplePayButtonProps = Omit<Config, 'supportedNetworks'> & {
     supportedNetworks?: CardNetwork[];
     transaction: Transaction;
     onDidAuthorizePayment?: (data: ApplePayResponse) => void;
