+++ packages/react-native-v2/src/index.ts
@@ -19,4 +19,4 @@ export type {
 
 export * from "./Card";
 export * from "./ThreeDSecure";
-export * from './ApplePayButton';
+export * from './ApplePayButton/index';
